requirements.txt
streamlit~=1.30.0
pandas~=2.1.4
numpy~=1.26.3
matplotlib~=3.8.2
plotly~=5.18.0
scikit-learn~=1.3.2
tensorflow~=2.15.0
nltk~=3.8.1
wordcloud~=1.9.3
transformers~=4.36.2
pillow~=10.2.0